+++ pdfalyzer/__init__.py
@@ -30,7 +30,8 @@ from pdfalyzer.helpers.rich_text_helper import print_highlighted
 from pdfalyzer.output.pdfalyzer_presenter import PdfalyzerPresenter
 from pdfalyzer.output.styles.rich_theme import PDFALYZER_THEME_DICT
 from pdfalyzer.pdfalyzer import Pdfalyzer
-from pdfalyzer.util.argument_parser import ask_to_proceed, output_sections, parse_arguments, parse_combine_pdfs_args
+from pdfalyzer.util.argument_parser import (MAX_QUALITY, ask_to_proceed, output_sections, parse_arguments,
+     parse_combine_pdfs_args)
 from pdfalyzer.util.pdf_parser_manager import PdfParserManager
 
 # For the table shown by running pdfalyzer_show_color_theme
@@ -108,16 +109,16 @@ def combine_pdfs():
             print_highlighted(f"      -> Failed to merge '{pdf}'! {e}", style='red')
             ask_to_proceed()
 
-    if args.compression_level == 0:
-        print_highlighted("\nSkipping content stream compression...")
-    else:
-        print_highlighted(f"\nCompressing content streams with zlib level {args.compression_level}...")
+    # Iterate through pages and compress, lowering image quality if requested
+    # See https://pypdf.readthedocs.io/en/latest/user/file-size.html#reducing-image-quality
+    for i, page in enumerate(merger.pages):
+        if args.image_quality < MAX_QUALITY:
+            for j, img in enumerate(page.images):
+                print_highlighted(f"  -> Reducing image #{j + 1} quality on page {i + 1} to {args.image_quality}...", style='dim')
+                img.replace(img.image, quality=args.image_quality)
 
-        for i, page in enumerate(merger.pages):
-            # TODO: enable image quality reduction + zlib level once PyPDF is upgraded to 4.x and option is available
-            # See https://pypdf.readthedocs.io/en/latest/user/file-size.html#reducing-image-quality
-            print_highlighted(f"  -> Compressing page {i + 1}...", style='dim')
-            page.compress_content_streams()  # This is CPU intensive!
+        print_highlighted(f"  -> Compressing page {i + 1}...", style='dim')
+        page.compress_content_streams()  # This is CPU intensive!
 
     print_highlighted(f"\nWriting '{args.output_file}'...", style='cyan')
     merger.compress_identical_objects(remove_identicals=True, remove_orphans=True)