+++ tests/test_combine_pdfs.py
@@ -36,3 +36,10 @@ def test_combine_pdfs(combined_pdf_path, one_page_pdfs):
     reval = check_output([COMBINE_PDFS, '-o', combined_pdf_path, *one_page_pdfs], env=environ).decode()
     assert combined_pdf_path.exists()
     assert file_size_in_mb(combined_pdf_path) == 0.18
+
+
+def test_combine_pdfs_image_quality(combined_pdf_path, one_page_pdfs):
+    assert not combined_pdf_path.exists()
+    reval = check_output([COMBINE_PDFS, '-o', combined_pdf_path, '-iq', '1', *one_page_pdfs], env=environ).decode()
+    assert combined_pdf_path.exists()
+    assert file_size_in_mb(combined_pdf_path) == 0.09